+++ tasq/__init__.py
@@ -4,6 +4,6 @@ from __future__ import absolute_import, division, print_function, unicode_litera
 
 
 from tasq.remote.client import TasqClient
-from tasq.remote.master import Master
+from tasq.remote.master import Master, Masters
 
-__version__ = '0.5.0'
+__version__ = '0.6.0'