+++ tasq/worker.py
@@ -16,7 +16,7 @@ import logging
 from multiprocessing import Process
 
 
-_formatter = logging.Formatter('%(levelname)s - %(processName)s - %(message)s', '%Y-%m-%d %H:%M:%S')
+_fmt = logging.Formatter('%(message)s', '%Y-%m-%d %H:%M:%S')
 
 
 class Worker(Process):
@@ -40,7 +40,7 @@ class Worker(Process):
         # Logging settings
         self._log = logging.getLogger(f'{__name__}.{self._name}')
         sh = logging.StreamHandler()
-        sh.setFormatter(_formatter)
+        sh.setFormatter(_fmt)
         if self._debug is True:
             sh.setLevel(logging.DEBUG)
             self._log.setLevel(logging.DEBUG)